util: Rename MR to MR map
