+++ Makefile.am
@@ -58,7 +58,7 @@ common_srcs =				\
 	prov/util/src/util_poll.c	\
 	prov/util/src/util_wait.c	\
 	prov/util/src/util_buf.c	\
-	prov/util/src/util_mr.c		\
+	prov/util/src/util_mr_map.c	\
 	prov/util/src/util_ns.c		\
 	prov/util/src/util_shm.c	\
 	prov/util/src/util_mem_monitor.c
